Add webgl map fallback
Add wrapper around the map component to display the same loader and
error messages that are displayed in the preview embed.  More work needs
to be done to completely reuse the same components. Will be done in a
future PR.

Fix #627.
