+++ __tests__/lib/browser/webgl.js
@@ -1,0 +1,7 @@
+import {isWebglSupported} from '../../../lib/browser/webgl'
+
+describe('isWebglSupported', () => {
+  it('should be false in test environment', () => {
+    expect(isWebglSupported()).toBe(false)
+  })
+})
