+++ locales/fr/common.json
@@ -70,6 +70,10 @@
   },
 
   "map": {
+    "loading": "Chargement de la carte",
+    "errors": {
+      "noWebgl": "Nous ne pouvons pas afficher cette carte car votre navigateur ne supporte pas WebGL"
+    },
     "additionalFeature": "Un objet de plus…",
     "additionalFeature_plural": "{{ count }} objets de plus…"
   },
