+++ locales/fr/preview.json
@@ -5,8 +5,7 @@
       "init": "Initialisation du téléchargement",
       "progress": "Téléchargement en cours : {{ bytes }}",
       "progressWithTotal": "Téléchargement en cours : {{ bytes }} / {{ total }}"
-    },
-    "component": "Chargement de la carte"
+    }
   },
 
   "error": {
